+++ expyriment/io/_touchscreenbuttonbox.py
@@ -68,8 +68,11 @@ class TouchScreenButtonBox(Input):
         self._button_fields = []
         self._stimuli = []
         self.background_stimulus = background_stimulus
-        map(self.add_button_field, button_fields)
-        map(self.add_stimulus, stimuli)
+
+        for x in button_fields:
+            self.add_button_field(x)
+        for x in stimuli:
+            self.add_stimulus(x)
 
     def add_button_field(self, button_field):
         """Add a touchscreen button fields.
@@ -80,6 +83,7 @@ class TouchScreenButtonBox(Input):
 
         """
 
+        print("HH")
         if not isinstance(button_field, stimuli._visual.Visual):
             raise TypeError("Button field has to be a visual Expyriment stimulus")
         self._button_fields.append(button_field)